add folders for CUNIT tests
