+++ tests/CUNIT/srcs/minishell.test.c
@@ -6,14 +6,11 @@
 /*   By: rbroque <rbroque@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/04/02 22:28:05 by rbroque           #+#    #+#             */
-/*   Updated: 2023/04/03 00:26:10 by rbroque          ###   ########.fr       */
+/*   Updated: 2023/04/03 00:51:28 by rbroque          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include <CUnit/CUnit.h>
-#include <CUnit/Basic.h>
-#include <stdbool.h>
-#include <stdlib.h>
+#include "cunit.test.h"
 
 int	maxi(const int a, const int b)
 {
@@ -33,21 +30,6 @@ int	mini(const int a, const int b)
 //////////// TEST UTILS ////////////
 ////////////////////////////////////
 
-#define MAX_TESTS_COUNT 10
-
-typedef struct s_test
-{
-	char	*tname;
-	void	(*fct)(void);
-}				t_test;
-
-typedef struct s_suite
-{
-	CU_pSuite suite;
-	char	*sname;
-	t_test	tests[MAX_TESTS_COUNT];
-}				t_suite;
-
 void	test_maxi(void)
 {
 	CU_ASSERT_FATAL(maxi(0, -2) == 0);
@@ -60,65 +42,6 @@ void	test_mini(void)
 	CU_ASSERT_FATAL(mini(0, -2) == -2);
 	CU_ASSERT_FATAL(mini(1, -2) == -2);
 	CU_ASSERT_FATAL(mini(0, 0) == 0);
-}
-
-bool	are_suites_allocated(t_suite array[])
-{
-	size_t	i;
-
-	i = 0;
-	while (array[i].sname != NULL && array[i].suite != NULL)
-		++i;
-	return (array[i].sname == NULL && array[i].suite == NULL);
-}
-
-void	exit_tests(void)
-{
-	CU_cleanup_registry();
-	exit(CU_get_error());
-}
-
-void	add_suites(t_suite suite_mapping[])
-{
-	size_t	i;
-
-	i = 0;
-	while (suite_mapping[i].sname != NULL)
-	{
-		suite_mapping[i].suite = CU_add_suite(suite_mapping[i].sname, NULL, NULL);
-		++i;
-	}
-}
-
-void	add_tests(t_suite suite_mapping[])
-{
-	size_t	i;
-	size_t	j;
-
-	i = 0;
-	while (suite_mapping[i].sname != NULL)
-	{
-		j = 0;
-		while (suite_mapping[i].tests[j].tname != NULL)
-		{
-			if (CU_add_test(
-					suite_mapping[i].suite,
-					suite_mapping[i].tests[j].tname,
-					suite_mapping[i].tests[j].fct)
-				== NULL)
-				exit_tests();
-			++j;
-		}
-		++i;
-	}
-}
-
-void	map_suite(t_suite suite_mapping[])
-{
-	add_suites(suite_mapping);
-	if (are_suites_allocated(suite_mapping) == false)
-		exit_tests();
-	add_tests(suite_mapping);
 }
 
 void	set_tests(void)
@@ -135,6 +58,10 @@ void	set_tests(void)
 		{.suite=NULL, .sname="hybride_suite", .tests={
 			{.tname="test_maxi", test_maxi},
 			{.tname="test_mini", test_mini},
+			}
+		},
+		{.suite=NULL, .sname="Quotes", .tests={
+			{.tname="are_quotes_closed", are_quotes_closed__test},
 			}
 		},
 		{.suite=NULL, .sname=NULL}
